Move GCP functions from console.utility to project.cloud.gcp.
These functions pertain to management of files in Google Cloud Storage
buckets as well as management of Google Cloud access groups.

In order to organize things slightly better, and remove lower-level
implementation code from the console app, move these functions into a
new module project.cloud.gcp.
